VCTCsag class and results/VTCDresults working. VTCDsagCircuit not working, but I am trying.
